Safe Alert Fixes 0905
Issue: https://gitcode.com/openharmony/arkcompiler_runtime_core/issues/12569

AI[22%] Human Fixed[0%] Human[78%] AI Adopted[100%]

+++ static_core/plugins/ets/runtime/intrinsics/std_core_gc_lifecycle.cpp
@@ -51,7 +51,7 @@ extern "C" EtsLong StdGCStartGC(EtsInt cause, EtsObject *callback, EtsBoolean is
     }
     auto &gcTaskTracker = GCTaskTracker::InitIfNeededAndGet(gc);
     auto task = MakePandaUnique<GCTask>(reason);
-    uint32_t id = task->GetId();
+    uint64_t id = task->GetId();
     if (callback != nullptr) {
         auto *callbackRef = executionCtx->GetPandaVM()->GetGlobalObjectStorage()->Add(
             callback->GetCoreType(), mem::Reference::ObjectType::GLOBAL);
